fix(conversation): #COCO-4203 refacto
