+++ conversation/frontend/src/features/message-edit/MessageEdit.tsx
@@ -1,24 +1,23 @@
-import { FormControl, Input, useDate, useDebounce } from '@edifice.io/react';
-import { useEffect, useRef, useState } from 'react';
+import { FormControl, Input } from '@edifice.io/react';
+import { useEffect, useState } from 'react';
 import { MessageActionDropdown } from '~/components/MessageActionDropdown/MessageActionDropdown';
 import { MessageBody } from '~/components/MessageBody';
 import { useI18n } from '~/hooks/useI18n';
 import { useMessageIdAndAction } from '~/hooks/useMessageIdAndAction';
 import { Message } from '~/models';
-import { useConversationConfig, useCreateOrUpdateDraft } from '~/services';
-import { useMessageActions, useMessageNeedToSave } from '~/store/messageStore';
+import { useCreateOrUpdateDraft } from '~/services';
+import { useMessageActions } from '~/store/messageStore';
 import { MessageEditHeader } from './components/MessageEditHeader';
+import { MessageSaveDate } from './components/MessageSaveDate';
+import { useAutoSaveMessage } from './hooks/useAutoSaveMessage';
 
 export function MessageEdit({ message }: { message?: Message }) {
+  console.log('message:', message);
   const { t } = useI18n();
   const [subject, setSubject] = useState(message?.subject);
-  const messageUpdatedNeedSave = useMessageNeedToSave();
   const { setMessage, setMessageNeedToSave } = useMessageActions();
-  const { fromNow } = useDate();
-  const debounceTimeToSave = useRef(3000);
   const createOrUpdateDraft = useCreateOrUpdateDraft();
-  const [dateKey, setDateKey] = useState(0);
-  const { data: publicConfig } = useConversationConfig();
+  useAutoSaveMessage();
   const { action } = useMessageIdAndAction();
   const isTransferAction = action === 'transfer';
 
@@ -34,11 +33,6 @@ export function MessageEdit({ message }: { message?: Message }) {
     setMessageNeedToSave(true);
   };
 
-  const messageUpdatedNeedSaveDebounced = useDebounce(
-    messageUpdatedNeedSave,
-    debounceTimeToSave.current,
-  );
-
   useEffect(() => {
     // Automatically create draft when this is a transfer action
     // so the attachments are transferred to the new message
@@ -46,26 +40,7 @@ export function MessageEdit({ message }: { message?: Message }) {
     if (message && !message.id && isTransferAction) {
       createOrUpdateDraft();
     }
-
-    const interval = setInterval(() => setDateKey((prev) => ++prev), 6000);
-
-    return () => clearInterval(interval);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
-
-  useEffect(() => {
-    if (publicConfig && publicConfig['debounce-time-to-auto-save']) {
-      debounceTimeToSave.current = publicConfig['debounce-time-to-auto-save'];
-    }
-  }, [publicConfig]);
-
-  useEffect(() => {
-    if (messageUpdatedNeedSaveDebounced) {
-      createOrUpdateDraft();
-      setMessageNeedToSave(false);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [messageUpdatedNeedSaveDebounced]);
 
   return (
     <>
@@ -98,11 +73,7 @@ export function MessageEdit({ message }: { message?: Message }) {
                 }}
                 className="gap-12"
               />
-              {message?.date && (
-                <div className="caption fst-italic" key={dateKey}>
-                  {t('message.saved') + ' ' + fromNow(message.date)}
-                </div>
-              )}
+              <MessageSaveDate />
             </div>
           </div>
         </div>
